# Optimize prosper-only build times