--- a/src/Window.cpp
+++ b/src/Window.cpp
@@ -1,5 +1,6 @@
 #include "Window.hpp"
 
+#include <GLFW/glfw3.h>
 #include <imgui.h>
 #include <imgui_impl_glfw.h>