--- a/src/gfx/DescriptorAllocator.cpp
+++ b/src/gfx/DescriptorAllocator.cpp
@@ -1,6 +1,7 @@
 #include "DescriptorAllocator.hpp"
 
 #include "../utils/Utils.hpp"
+#include "Device.hpp"
 #include "VkUtils.hpp"
 
 using namespace wheels;