--- a/src/gfx/ShaderReflection.hpp
+++ b/src/gfx/ShaderReflection.hpp
@@ -6,7 +6,6 @@
 #include <wheels/containers/hash_map.hpp>
 #include <wheels/containers/hash_set.hpp>
 #include <wheels/containers/span.hpp>
-#include <wheels/containers/static_array.hpp>
 #include <wheels/containers/string.hpp>
 
 #include <vulkan/vulkan.hpp>
@@ -64,11 +63,11 @@ class ShaderReflection
 
     // TODO: This doesn't deduce N from infos defined as an initializer list
     // Takes bindings sorted by the glsl binding indices
-    template <size_t N>
-    [[nodiscard]] wheels::StaticArray<vk::WriteDescriptorSet, N>
+    [[nodiscard]] wheels::Array<vk::WriteDescriptorSet>
     generateDescriptorWrites(
-        uint32_t descriptorSetIndex, vk::DescriptorSet descriptorSetHandle,
-        const wheels::StaticArray<DescriptorInfo, N> &descriptorInfos) const;
+        wheels::Allocator &alloc, uint32_t descriptorSetIndex,
+        vk::DescriptorSet descriptorSetHandle,
+        wheels::Span<const DescriptorInfo> descriptorInfos) const;
 
   private:
     uint32_t _pushConstantsBytesize{0};
@@ -76,69 +75,5 @@ class ShaderReflection
         _descriptorSetMetadatas;
     wheels::HashSet<std::filesystem::path> _sourceFiles;
 };
-
-template <size_t N>
-wheels::StaticArray<vk::WriteDescriptorSet, N> ShaderReflection::
-    generateDescriptorWrites(
-        uint32_t descriptorSetIndex, vk::DescriptorSet descriptorSetHandle,
-        const wheels::StaticArray<DescriptorInfo, N> &descriptorInfos) const
-{
-    const wheels::Array<DescriptorSetMetadata> *metadatas =
-        _descriptorSetMetadatas.find(descriptorSetIndex);
-    WHEELS_ASSERT(metadatas != nullptr);
-    // false positive, custom assert above
-    // NOLINTNEXTLINE(clang-analyzer-core.CallAndMessage)
-    WHEELS_ASSERT(metadatas->size() == N);
-
-    wheels::StaticArray<vk::WriteDescriptorSet, N> descriptorWrites;
-    WHEELS_ASSERT(descriptorInfos.size() == N);
-
-    for (uint32_t i = 0; i < N; ++i)
-    {
-        const DescriptorInfo &descriptorInfo = descriptorInfos[i];
-
-        const vk::DescriptorImageInfo *pImageInfo =
-            std::get_if<vk::DescriptorImageInfo>(&descriptorInfo);
-        const vk::DescriptorBufferInfo *pBufferInfo =
-            std::get_if<vk::DescriptorBufferInfo>(&descriptorInfo);
-        const vk::BufferView *pTexelBufferView =
-            std::get_if<vk::BufferView>(&descriptorInfo);
-        // TODO:
-        // Refactor this so that single image is also a span? How are the
-        // ergonomics?
-        const wheels::Span<const vk::DescriptorImageInfo> *pImageInfoSpan =
-            std::get_if<wheels::Span<const vk::DescriptorImageInfo>>(
-                &descriptorInfo);
-        const wheels::Span<const vk::DescriptorBufferInfo> *pBufferInfoSpan =
-            std::get_if<wheels::Span<const vk::DescriptorBufferInfo>>(
-                &descriptorInfo);
-
-        uint32_t descriptorCount = 1;
-
-        if (pImageInfoSpan != nullptr)
-        {
-            pImageInfo = pImageInfoSpan->data();
-            descriptorCount = asserted_cast<uint32_t>(pImageInfoSpan->size());
-        }
-        else if (pBufferInfoSpan != nullptr)
-        {
-            pBufferInfo = pBufferInfoSpan->data();
-            descriptorCount = asserted_cast<uint32_t>(pBufferInfoSpan->size());
-        }
-
-        const DescriptorSetMetadata &metadata = (*metadatas)[i];
-        descriptorWrites.push_back(vk::WriteDescriptorSet{
-            .dstSet = descriptorSetHandle,
-            .dstBinding = metadata.binding,
-            .descriptorCount = descriptorCount,
-            .descriptorType = metadata.descriptorType,
-            .pImageInfo = pImageInfo,
-            .pBufferInfo = pBufferInfo,
-            .pTexelBufferView = pTexelBufferView,
-        });
-    }
-
-    return descriptorWrites;
-}
 
 #endif // PROSPER_GFX_SHADER_REFLECTION_HPP