--- a/src/render/ComputePass.cpp
+++ b/src/render/ComputePass.cpp
@@ -1,6 +1,6 @@
 #include "ComputePass.hpp"
 
-#include <glm/glm.hpp>
+#include <glm/vector_relational.hpp>
 #include <imgui.h>
 
 #include "../gfx/DescriptorAllocator.hpp"