Make compilation faster with using variety of tactics for the exercise. This drops total build time on my w10 5900x box from 6.8s down to 6.2s and speeds up some CUs by half. 6850u linux clang build gets from 17s down to 13s with similar CU improvements.

The improvements are a drop in the bucket for a full rebuild but the snappier per-file rebuilds are felt in normal dev. Dependency build optimization is a whole different ball game and doesn't seem worth the extra time, not to mention the extra maintenance overhead for new versions.